+++ duoStrength.js
@@ -135,6 +135,7 @@ function retrieveOptions()
 						"skillSuggestionMethod":					"0",
 						"hideSuggestionNonStrengthened":			true,
 						"hideSuggestionWithCrackedSkills":			true,
+						"suggestionPopoutButton":					true,
 					"focusFirstSkill":							true,
 					"practiseButton":							true,
 					"practiceType":								"0",
@@ -423,6 +424,15 @@ function removeFocusModeButton()
 	if (focusModeButton !== null)
 	{
 		focusModeButton.remove();
+	}
+}
+
+function removeSuggestionPopoutButton()
+{
+	const suggestionPopoutButton = document.querySelector(`#suggestionPopoutButton`);
+	if (suggestionPopoutButton !== null)
+	{
+		suggestionPopoutButton.remove();
 	}
 }
 
@@ -2305,6 +2315,43 @@ function getLanguagesInfo()
 	return languagesInfo;
 }
 
+function addSuggestionPopoutButton(skillName)
+{
+	const skillElement = Array.from(document.querySelectorAll(`${SKILL_SELECTOR}`)).find(
+		(skill) => 
+		{
+			return skill.querySelector(SKILL_NAME_SELECTOR).textContent === skillName;
+		}
+	);
+	
+	const button = document.createElement("button");
+	button.title = `Open popout for ${skillName}`;
+	button.id = "suggestionPopoutButton";
+	button.addEventListener("click",
+		(event) =>
+		{
+			event.stopPropagation();
+			skillElement.firstChild.click();
+		}
+	);
+	button.style =
+	`
+		background: none;
+		border: none;
+		line-height: 1em;
+		cursor: pointer;
+	`;
+	button.appendChild(document.createElement("img"));
+	button.lastChild.src = chrome.runtime.getURL("images/popout.svg");
+	button.lastChild.style =
+	`
+		width: 1em;
+		vertical-align: middle;
+	`;
+
+	return button;
+}
+
 function displayCrownsBreakdown()
 {
 	if (Object.entries(userData).length === 0)
@@ -3195,7 +3242,7 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 				break;
 		}
 
-		let link = document.createElement("a");
+		const link = document.createElement("a");
 		link.href = "/skill/" + languageCode + "/" + randomSuggestion.url_title + ((treeLevel == 5) ? "/practice/" : "/");
 		link.textContent = randomSuggestion.short;
 		link.style.color = 'blue';
@@ -3215,7 +3262,7 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 			}
 		);
 		
-		let suggestionMessage = document.createElement("p");
+		const suggestionMessage = document.createElement("p");
 		if (treeLevel == 5)
 		{
 			let messageText = `Your ${language} tree is `
@@ -3277,6 +3324,7 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 		}
 		else
 		{
+			// Not level 5 or level 0 so we will use the suggestion based on the users options.
 			if (fullyStrengthened && noCrackedSkills)
 				suggestionMessage.textContent = `Your ${language} tree is fully strengthened, and there are no cracked cracked skills. `;
 			else if (fullyStrengthened)
@@ -3288,6 +3336,7 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 			suggestionMessage.appendChild(link);
 		}
 
+
 		container.appendChild(suggestionMessage);
 
 		topOfTree.appendChild(container);
@@ -3298,6 +3347,18 @@ function displaySuggestion(fullyStrengthened, noCrackedSkills)
 	{
 		// Already made the box.
 		// And we don't want to do anything otherwise the suggestion, if using the random option, will keep changing and might look a bit strange.
+	}
+
+	// Skill Suggestion Popout Button
+	removeSuggestionPopoutButton(); // Remove if already exists.
+
+	const suggestionName = document.querySelector(`#fullStrengthMessageContainer a`).textContent;
+
+	if (options.suggestionPopoutButton && Array.from(document.querySelectorAll(SKILL_NAME_SELECTOR)).some(skillName => skillName.textContent === suggestionName))
+	{
+		// Add button that opens up the suggested skill's popout
+		const button = addSuggestionPopoutButton(suggestionName);
+		document.querySelector(`#fullStrengthMessageContainer p`).appendChild(button);
 	}
 }